+++ rust/cubesqlplanner/cubesqlplanner/src/plan/builder/select.rs
@@ -1,27 +1,30 @@
 use crate::plan::{
-    AliasedExpr, Cte, Expr, Filter, From, MemberExpression, OrderBy, Schema, Select,
+    AliasedExpr, Cte, Expr, Filter, From, MemberExpression, OrderBy, QualifiedColumnName, Schema,
+    SchemaColumn, Select, SingleAliasedSource, SingleSource,
 };
+
+use crate::plan::expression::FunctionExpression;
+use crate::planner::sql_evaluator::sql_nodes::SqlNodesFactory;
 use crate::planner::{BaseMember, VisitorContext};
+use std::collections::HashMap;
 use std::rc::Rc;
 
 pub struct SelectBuilder {
     projection_columns: Vec<AliasedExpr>,
-    from: From,
+    from: Rc<From>,
     filter: Option<Filter>,
     group_by: Vec<Expr>,
     having: Option<Filter>,
     order_by: Vec<OrderBy>,
-    context: Rc<VisitorContext>,
     ctes: Vec<Rc<Cte>>,
     is_distinct: bool,
     limit: Option<usize>,
     offset: Option<usize>,
-    input_schema: Rc<Schema>,
+    result_schema: Schema,
 }
 
 impl SelectBuilder {
-    pub fn new(from: From, context: VisitorContext) -> Self {
-        let input_schema = from.schema.clone();
+    pub fn new(from: Rc<From>) -> Self {
         Self {
             projection_columns: vec![],
             from,
@@ -29,33 +32,60 @@ impl SelectBuilder {
             group_by: vec![],
             having: None,
             order_by: vec![],
-            context: Rc::new(context),
             ctes: vec![],
             is_distinct: false,
             limit: None,
             offset: None,
-            input_schema,
+            result_schema: Schema::empty(),
         }
     }
 
-    pub fn add_projection_member(
-        &mut self,
-        member: &Rc<dyn BaseMember>,
-        source: Option<String>,
-        alias: Option<String>,
-    ) {
+    pub fn add_projection_member(&mut self, member: &Rc<dyn BaseMember>, alias: Option<String>) {
         let alias = if let Some(alias) = alias {
             alias
         } else {
-            self.input_schema.resolve_member_alias(&member, &source)
+            member.alias_name()
         };
-        let expr = Expr::Member(MemberExpression::new(member.clone(), source));
+
+        let expr = Expr::Member(MemberExpression::new(member.clone()));
         let aliased_expr = AliasedExpr {
             expr,
             alias: alias.clone(),
         };
 
         self.projection_columns.push(aliased_expr);
+        self.result_schema
+            .add_column(SchemaColumn::new(alias.clone(), Some(member.full_name())));
+    }
+
+    pub fn add_projection_coalesce_member(
+        &mut self,
+        member: &Rc<dyn BaseMember>,
+        references: Vec<QualifiedColumnName>,
+        alias: Option<String>,
+    ) {
+        let alias = if let Some(alias) = alias {
+            alias
+        } else {
+            member.alias_name()
+        };
+
+        let expr = Expr::Function(FunctionExpression {
+            function: "COALESCE".to_string(),
+            arguments: references
+                .into_iter()
+                // TODO unwrap
+                .map(|r| Expr::Reference(r))
+                .collect(),
+        });
+        let aliased_expr = AliasedExpr {
+            expr,
+            alias: alias.clone(),
+        };
+
+        self.projection_columns.push(aliased_expr);
+        self.result_schema
+            .add_column(SchemaColumn::new(alias.clone(), Some(member.full_name())));
     }
 
     pub fn set_filter(&mut self, filter: Option<Filter>) {
@@ -85,11 +115,65 @@ impl SelectBuilder {
     pub fn set_offset(&mut self, offset: Option<usize>) {
         self.offset = offset;
     }
+
     pub fn set_ctes(&mut self, ctes: Vec<Rc<Cte>>) {
         self.ctes = ctes;
     }
 
-    pub fn build(self) -> Select {
+    fn make_cube_references(&self) -> HashMap<String, String> {
+        let mut refs = HashMap::new();
+        match &self.from.source {
+            crate::plan::FromSource::Single(source) => {
+                self.add_cube_reference_if_needed(source, &mut refs)
+            }
+            crate::plan::FromSource::Join(join) => {
+                self.add_cube_reference_if_needed(&join.root, &mut refs);
+                for join_item in join.joins.iter() {
+                    self.add_cube_reference_if_needed(&join_item.from, &mut refs);
+                }
+            }
+            crate::plan::FromSource::Empty => {}
+        }
+        refs
+    }
+
+    fn add_cube_reference_if_needed(
+        &self,
+        source: &SingleAliasedSource,
+        refs: &mut HashMap<String, String>,
+    ) {
+        match &source.source {
+            SingleSource::Cube(cube) => {
+                refs.insert(cube.name().clone(), source.alias.clone());
+            }
+            _ => {}
+        }
+    }
+
+    fn make_asteriks_schema(&self) -> Rc<Schema> {
+        let schema = match &self.from.source {
+            crate::plan::FromSource::Empty => Rc::new(Schema::empty()),
+            crate::plan::FromSource::Single(source) => source.source.schema(),
+            crate::plan::FromSource::Join(join) => {
+                let mut schema = Schema::empty();
+                schema.merge(join.root.source.schema().as_ref());
+                for itm in join.joins.iter() {
+                    schema.merge(itm.from.source.schema().as_ref())
+                }
+                Rc::new(schema)
+            }
+        };
+        schema
+    }
+
+    pub fn build(self, mut nodes_factory: SqlNodesFactory) -> Select {
+        let cube_references = self.make_cube_references();
+        nodes_factory.set_cube_name_references(cube_references);
+        let schema = if self.projection_columns.is_empty() {
+            self.make_asteriks_schema()
+        } else {
+            Rc::new(self.result_schema)
+        };
         Select {
             projection_columns: self.projection_columns,
             from: self.from,
@@ -97,11 +181,12 @@ impl SelectBuilder {
             group_by: self.group_by,
             having: self.having,
             order_by: self.order_by,
-            context: self.context.clone(),
+            context: Rc::new(VisitorContext::new(&nodes_factory)),
             ctes: self.ctes,
             is_distinct: self.is_distinct,
             limit: self.limit,
             offset: self.offset,
+            schema,
         }
     }
 }