+++ rust/cubesqlplanner/cubesqlplanner/src/plan/expression.rs
@@ -1,4 +1,4 @@
-use super::Schema;
+use super::QualifiedColumnName;
 use crate::planner::sql_templates::PlanSqlTemplates;
 use crate::planner::{BaseMember, VisitorContext};
 use cubenativeutils::CubeError;
@@ -7,44 +7,64 @@ use std::rc::Rc;
 #[derive(Clone)]
 pub struct MemberExpression {
     pub member: Rc<dyn BaseMember>,
-    pub source: Option<String>,
 }
 
 impl MemberExpression {
-    pub fn new(member: Rc<dyn BaseMember>, source: Option<String>) -> Self {
-        Self { member, source }
+    pub fn new(member: Rc<dyn BaseMember>) -> Self {
+        Self { member }
     }
 
     pub fn to_sql(
         &self,
-        templates: &PlanSqlTemplates,
+        _templates: &PlanSqlTemplates,
         context: Rc<VisitorContext>,
-        schema: Rc<Schema>,
     ) -> Result<String, CubeError> {
-        if let Some(reference_column) =
-            schema.find_column_for_member(&self.member.full_name(), &self.source)
-        {
-            templates.column_reference(&reference_column.table_name, &reference_column.alias)
-        } else {
-            self.member.to_sql(context, schema)
-        }
+        self.member.to_sql(context)
     }
+}
+
+#[derive(Clone)]
+pub struct FunctionExpression {
+    pub function: String,
+    pub arguments: Vec<Expr>,
 }
 
 #[derive(Clone)]
 pub enum Expr {
     Member(MemberExpression),
+    Reference(QualifiedColumnName),
+    Function(FunctionExpression),
 }
 
 impl Expr {
+    pub fn new_member(member: Rc<dyn BaseMember>) -> Self {
+        Self::Member(MemberExpression::new(member))
+    }
+    pub fn new_reference(source: Option<String>, reference: String) -> Self {
+        Self::Reference(QualifiedColumnName::new(source, reference))
+    }
     pub fn to_sql(
         &self,
         templates: &PlanSqlTemplates,
         context: Rc<VisitorContext>,
-        schema: Rc<Schema>,
     ) -> Result<String, CubeError> {
         match self {
-            Expr::Member(member) => member.to_sql(templates, context, schema),
+            Self::Member(member) => member.to_sql(templates, context),
+            Self::Reference(reference) => {
+                templates.column_reference(reference.source(), &reference.name())
+            }
+            Expr::Function(FunctionExpression {
+                function,
+                arguments,
+            }) => templates.scalar_function(
+                function.to_string(),
+                arguments
+                    .iter()
+                    .map(|e| e.to_sql(&templates, context.clone()))
+                    .collect::<Result<Vec<_>, _>>()?,
+                None,
+                None,
+            ),
         }
     }
 }