+++ rust/cubesqlplanner/cubesqlplanner/src/plan/union.rs
@@ -1,21 +1,25 @@
 use super::{QueryPlan, Schema};
 use crate::planner::sql_templates::PlanSqlTemplates;
 use cubenativeutils::CubeError;
+use std::rc::Rc;
 
 pub struct Union {
     pub union: Vec<QueryPlan>,
+    pub schema: Rc<Schema>,
 }
 
 impl Union {
     pub fn new(union: Vec<QueryPlan>) -> Self {
-        Self { union }
-    }
-    pub fn make_schema(&self, self_alias: Option<String>) -> Schema {
-        if self.union.is_empty() {
-            Schema::empty()
+        let schema = if union.is_empty() {
+            Rc::new(Schema::empty())
         } else {
-            self.union[0].make_schema(self_alias)
-        }
+            union[0].schema()
+        };
+        Self { union, schema }
+    }
+
+    pub fn schema(&self) -> Rc<Schema> {
+        self.schema.clone()
     }
 
     pub fn to_sql(&self, templates: &PlanSqlTemplates) -> Result<String, CubeError> {