+++ rust/cubesqlplanner/cubesqlplanner/src/planner/base_cube.rs
@@ -1,40 +1,52 @@
 use super::query_tools::QueryTools;
-use super::sql_evaluator::EvaluationNode;
+use super::sql_evaluator::MemberSymbol;
 use super::{evaluate_with_context, VisitorContext};
-use crate::plan::Schema;
 use cubenativeutils::CubeError;
+use std::collections::HashSet;
 use std::rc::Rc;
 
 pub struct BaseCube {
     cube_name: String,
-    member_evaluator: Rc<EvaluationNode>,
+    members: HashSet<String>,
+    member_evaluator: Rc<MemberSymbol>,
     query_tools: Rc<QueryTools>,
 }
 impl BaseCube {
     pub fn try_new(
         cube_name: String,
         query_tools: Rc<QueryTools>,
-        member_evaluator: Rc<EvaluationNode>,
+        member_evaluator: Rc<MemberSymbol>,
     ) -> Result<Rc<Self>, CubeError> {
+        let members = query_tools
+            .base_tools()
+            .all_cube_members(cube_name.clone())?
+            .into_iter()
+            .collect::<HashSet<_>>();
+
         Ok(Rc::new(Self {
             cube_name,
+            members,
             member_evaluator,
             query_tools,
         }))
     }
 
     pub fn to_sql(&self, context: Rc<VisitorContext>) -> Result<String, CubeError> {
-        let cube_sql = evaluate_with_context(
-            &self.member_evaluator,
-            self.query_tools.clone(),
-            context,
-            Rc::new(Schema::empty()),
-        )?;
+        let cube_sql =
+            evaluate_with_context(&self.member_evaluator, self.query_tools.clone(), context)?;
         Ok(cube_sql)
     }
 
     pub fn name(&self) -> &String {
         &self.cube_name
+    }
+
+    pub fn members(&self) -> &HashSet<String> {
+        &self.members
+    }
+
+    pub fn has_member(&self, name: &str) -> bool {
+        self.members.contains(name)
     }
 
     pub fn default_alias(&self) -> String {