+++ rust/cubesqlplanner/cubesqlplanner/src/planner/base_join_condition.rs
@@ -1,35 +1,29 @@
 use super::query_tools::QueryTools;
-use super::sql_evaluator::EvaluationNode;
-use super::{evaluate_with_context, VisitorContext};
-use crate::plan::Schema;
+use super::sql_evaluator::SqlCall;
+use super::{evaluate_sql_call_with_context, VisitorContext};
 use cubenativeutils::CubeError;
 use std::rc::Rc;
 pub trait BaseJoinCondition {
-    fn to_sql(&self, context: Rc<VisitorContext>, schema: Rc<Schema>) -> Result<String, CubeError>;
+    fn to_sql(&self, context: Rc<VisitorContext>) -> Result<String, CubeError>;
 }
 pub struct SqlJoinCondition {
-    member_evaluator: Rc<EvaluationNode>,
+    sql_call: Rc<SqlCall>,
     query_tools: Rc<QueryTools>,
 }
 impl SqlJoinCondition {
     pub fn try_new(
         query_tools: Rc<QueryTools>,
-        member_evaluator: Rc<EvaluationNode>,
+        sql_call: Rc<SqlCall>,
     ) -> Result<Rc<Self>, CubeError> {
         Ok(Rc::new(Self {
-            member_evaluator,
+            sql_call,
             query_tools,
         }))
     }
 }
 
 impl BaseJoinCondition for SqlJoinCondition {
-    fn to_sql(&self, context: Rc<VisitorContext>, schema: Rc<Schema>) -> Result<String, CubeError> {
-        evaluate_with_context(
-            &self.member_evaluator,
-            self.query_tools.clone(),
-            context,
-            schema,
-        )
+    fn to_sql(&self, context: Rc<VisitorContext>) -> Result<String, CubeError> {
+        evaluate_sql_call_with_context(&self.sql_call, self.query_tools.clone(), context)
     }
 }