+++ rust/cubesqlplanner/cubesqlplanner/src/planner/base_time_dimension.rs
@@ -1,8 +1,7 @@
 use super::query_tools::QueryTools;
-use super::sql_evaluator::EvaluationNode;
+use super::sql_evaluator::MemberSymbol;
 use super::BaseDimension;
 use super::{BaseMember, VisitorContext};
-use crate::plan::Schema;
 use cubenativeutils::CubeError;
 use std::rc::Rc;
 
@@ -11,34 +10,19 @@ pub struct BaseTimeDimension {
     query_tools: Rc<QueryTools>,
     granularity: Option<String>,
     date_range: Option<Vec<String>>,
+    alias_suffix: String,
 }
 
 impl BaseMember for BaseTimeDimension {
-    fn to_sql(
-        &self,
-        context: Rc<VisitorContext>,
-        source_schema: Rc<Schema>,
-    ) -> Result<String, CubeError> {
-        let field_sql = if let Some(granularity) = &self.granularity {
-            let converted_tz = self
-                .query_tools
-                .base_tools()
-                .convert_tz(self.dimension.to_sql(context, source_schema)?)?;
-            self.query_tools
-                .base_tools()
-                .time_grouped_column(granularity.clone(), converted_tz)?
-        } else {
-            unimplemented!("Time dimensions without granularity not supported yet")
-        };
-        Ok(field_sql)
+    fn to_sql(&self, context: Rc<VisitorContext>) -> Result<String, CubeError> {
+        self.dimension.to_sql(context)
     }
 
     fn alias_name(&self) -> String {
-        self.query_tools
-            .escape_column_name(&self.unescaped_alias_name())
+        self.unescaped_alias_name()
     }
 
-    fn member_evaluator(&self) -> Rc<EvaluationNode> {
+    fn member_evaluator(&self) -> Rc<MemberSymbol> {
         self.dimension.member_evaluator()
     }
 
@@ -55,28 +39,39 @@ impl BaseMember for BaseTimeDimension {
     }
 
     fn alias_suffix(&self) -> Option<String> {
-        let granularity = if let Some(granularity) = &self.granularity {
-            granularity
-        } else {
-            "day"
-        };
-        Some(granularity.to_string())
+        Some(self.alias_suffix.clone())
     }
 }
 
 impl BaseTimeDimension {
     pub fn try_new_required(
         query_tools: Rc<QueryTools>,
-        member_evaluator: Rc<EvaluationNode>,
+        member_evaluator: Rc<MemberSymbol>,
         granularity: Option<String>,
         date_range: Option<Vec<String>>,
     ) -> Result<Rc<Self>, CubeError> {
+        let alias_suffix = if let Some(granularity) = &granularity {
+            granularity.clone()
+        } else {
+            "day".to_string()
+        };
         Ok(Rc::new(Self {
             dimension: BaseDimension::try_new_required(member_evaluator, query_tools.clone())?,
             query_tools,
             granularity,
             date_range,
+            alias_suffix,
         }))
+    }
+
+    pub fn change_granularity(&self, new_granularity: Option<String>) -> Rc<Self> {
+        Rc::new(Self {
+            dimension: self.dimension.clone(),
+            query_tools: self.query_tools.clone(),
+            granularity: new_granularity,
+            date_range: self.date_range.clone(),
+            alias_suffix: self.alias_suffix.clone(),
+        })
     }
 
     pub fn get_granularity(&self) -> Option<String> {
@@ -95,7 +90,7 @@ impl BaseTimeDimension {
         self.dimension.clone()
     }
 
-    pub fn member_evaluator(&self) -> Rc<EvaluationNode> {
+    pub fn member_evaluator(&self) -> Rc<MemberSymbol> {
         self.dimension.member_evaluator()
     }