+++ rust/cubesqlplanner/cubesqlplanner/src/planner/planners/join_planner.rs
@@ -1,8 +1,9 @@
 use super::CommonUtils;
+use crate::cube_bridge::join_definition::JoinDefinition;
 use crate::cube_bridge::memeber_sql::MemberSql;
 use crate::plan::{From, JoinBuilder, JoinCondition};
 use crate::planner::query_tools::QueryTools;
-use crate::planner::sql_evaluator::EvaluationNode;
+use crate::planner::sql_evaluator::SqlCall;
 use crate::planner::SqlJoinCondition;
 use cubenativeutils::CubeError;
 use std::rc::Rc;
@@ -20,11 +21,20 @@ impl JoinPlanner {
         }
     }
 
-    pub fn make_join_node_with_prefix(
+    pub fn make_join_node_with_prefix_and_join_hints(
         &self,
         alias_prefix: &Option<String>, /*TODO dimensions for subqueries*/
-    ) -> Result<From, CubeError> {
-        let join = self.query_tools.cached_data().join()?.clone();
+        join_hints: Vec<String>,
+    ) -> Result<Rc<From>, CubeError> {
+        let join = self.query_tools.join_graph().build_join(join_hints)?;
+        self.make_join_node_impl(alias_prefix, join)
+    }
+
+    pub fn make_join_node_impl(
+        &self,
+        alias_prefix: &Option<String>,
+        join: Rc<dyn JoinDefinition>,
+    ) -> Result<Rc<From>, CubeError> {
         let root = self.utils.cube_from_path(join.static_data().root.clone())?;
         let joins = join.joins()?;
         if joins.items().is_empty() {
@@ -36,11 +46,11 @@ impl JoinPlanner {
             );
             for join in joins.items().iter() {
                 let definition = join.join()?;
-                let evaluator = self
+                let sql_call = self
                     .compile_join_condition(&join.static_data().original_from, definition.sql()?)?;
                 let on = JoinCondition::new_base_join(SqlJoinCondition::try_new(
                     self.query_tools.clone(),
-                    evaluator,
+                    sql_call,
                 )?);
                 let cube = self
                     .utils
@@ -56,17 +66,13 @@ impl JoinPlanner {
         }
     }
 
-    pub fn make_join_node(&self) -> Result<From, CubeError> {
-        self.make_join_node_with_prefix(&None)
-    }
-
     fn compile_join_condition(
         &self,
         cube_name: &String,
         sql: Rc<dyn MemberSql>,
-    ) -> Result<Rc<EvaluationNode>, CubeError> {
+    ) -> Result<Rc<SqlCall>, CubeError> {
         let evaluator_compiler_cell = self.query_tools.evaluator_compiler().clone();
         let mut evaluator_compiler = evaluator_compiler_cell.borrow_mut();
-        evaluator_compiler.add_join_condition_evaluator(cube_name.clone(), sql)
+        evaluator_compiler.compile_sql_call(&cube_name, sql)
     }
 }