+++ rust/cubesqlplanner/cubesqlplanner/src/planner/sql_evaluator/sql_node_transformers/mod.rs
@@ -1,3 +1,0 @@
-pub mod set_schema;
-
-pub use set_schema::set_schema;