+++ rust/cubesqlplanner/cubesqlplanner/src/planner/sql_evaluator/sql_nodes/evaluate_sql.rs
@@ -1,7 +1,7 @@
 use super::SqlNode;
 use crate::planner::query_tools::QueryTools;
+use crate::planner::sql_evaluator::MemberSymbol;
 use crate::planner::sql_evaluator::SqlEvaluatorVisitor;
-use crate::planner::sql_evaluator::{EvaluationNode, MemberSymbolType};
 use cubenativeutils::CubeError;
 use std::any::Any;
 use std::rc::Rc;
@@ -17,18 +17,22 @@ impl EvaluateSqlNode {
 impl SqlNode for EvaluateSqlNode {
     fn to_sql(
         &self,
-        visitor: &mut SqlEvaluatorVisitor,
-        node: &Rc<EvaluationNode>,
-        _query_tools: Rc<QueryTools>,
+        visitor: &SqlEvaluatorVisitor,
+        node: &Rc<MemberSymbol>,
+        query_tools: Rc<QueryTools>,
         node_processor: Rc<dyn SqlNode>,
     ) -> Result<String, CubeError> {
-        let args = visitor.evaluate_deps(node, node_processor.clone())?;
-        match node.symbol() {
-            MemberSymbolType::Dimension(ev) => ev.evaluate_sql(args),
-            MemberSymbolType::Measure(ev) => ev.evaluate_sql(args),
-            MemberSymbolType::CubeTable(ev) => ev.evaluate_sql(args),
-            MemberSymbolType::CubeName(ev) => ev.evaluate_sql(args),
-            MemberSymbolType::SimpleSql(ev) => ev.evaluate_sql(args),
+        match node.as_ref() {
+            MemberSymbol::Dimension(ev) => {
+                ev.evaluate_sql(visitor, node_processor.clone(), query_tools.clone())
+            }
+            MemberSymbol::Measure(ev) => {
+                ev.evaluate_sql(visitor, node_processor.clone(), query_tools.clone())
+            }
+            MemberSymbol::CubeTable(ev) => {
+                ev.evaluate_sql(visitor, node_processor.clone(), query_tools.clone())
+            }
+            MemberSymbol::CubeName(ev) => ev.evaluate_sql(),
         }
     }