+++ rust/cubesqlplanner/cubesqlplanner/src/planner/sql_evaluator/symbols/measure_symbol.rs
@@ -1,26 +1,30 @@
 use super::{MemberSymbol, SymbolFactory};
 use crate::cube_bridge::evaluator::CubeEvaluator;
-use crate::cube_bridge::measure_definition::{MeasureDefinition, TimeShiftReference};
-use crate::cube_bridge::memeber_sql::{MemberSql, MemberSqlArg};
-use crate::planner::sql_evaluator::{Compiler, Dependency, EvaluationNode};
+use crate::cube_bridge::measure_definition::{
+    MeasureDefinition, RollingWindow, TimeShiftReference,
+};
+use crate::cube_bridge::memeber_sql::MemberSql;
+use crate::planner::query_tools::QueryTools;
+use crate::planner::sql_evaluator::{sql_nodes::SqlNode, Compiler, SqlCall, SqlEvaluatorVisitor};
 use cubenativeutils::CubeError;
 use std::rc::Rc;
 
+#[derive(Clone)]
 pub struct MeasureOrderBy {
-    evaluation_node: Rc<EvaluationNode>,
+    sql_call: Rc<SqlCall>,
     direction: String,
 }
 
 impl MeasureOrderBy {
-    pub fn new(evaluation_node: Rc<EvaluationNode>, direction: String) -> Self {
+    pub fn new(sql_call: Rc<SqlCall>, direction: String) -> Self {
         Self {
-            evaluation_node,
+            sql_call,
             direction,
         }
     }
 
-    pub fn evaluation_node(&self) -> &Rc<EvaluationNode> {
-        &self.evaluation_node
+    pub fn sql_call(&self) -> &Rc<SqlCall> {
+        &self.sql_call
     }
 
     pub fn direction(&self) -> &String {
@@ -28,22 +32,24 @@ impl MeasureOrderBy {
     }
 }
 
+#[derive(Clone)]
 pub struct MeasureSymbol {
     cube_name: String,
     name: String,
     definition: Rc<dyn MeasureDefinition>,
-    measure_filters: Vec<Rc<EvaluationNode>>,
+    measure_filters: Vec<Rc<SqlCall>>,
     measure_order_by: Vec<MeasureOrderBy>,
-    member_sql: Rc<dyn MemberSql>,
+    member_sql: Rc<SqlCall>,
+    is_splitted_source: bool,
 }
 
 impl MeasureSymbol {
     pub fn new(
         cube_name: String,
         name: String,
-        member_sql: Rc<dyn MemberSql>,
+        member_sql: Rc<SqlCall>,
         definition: Rc<dyn MeasureDefinition>,
-        measure_filters: Vec<Rc<EvaluationNode>>,
+        measure_filters: Vec<Rc<SqlCall>>,
         measure_order_by: Vec<MeasureOrderBy>,
     ) -> Self {
         Self {
@@ -53,11 +59,30 @@ impl MeasureSymbol {
             definition,
             measure_filters,
             measure_order_by,
+            is_splitted_source: false,
         }
     }
 
     pub fn full_name(&self) -> String {
         format!("{}.{}", self.cube_name, self.name)
+    }
+
+    pub fn is_splitted_source(&self) -> bool {
+        self.is_splitted_source
+    }
+
+    pub fn split_with_source(&self, source_name: String) -> (Self, Self) {
+        let mut measure_with_source = self.clone();
+        measure_with_source.is_splitted_source = true;
+        let source = Self::new(
+            self.cube_name.clone(),
+            source_name.clone(),
+            self.member_sql.clone(),
+            self.definition().clone(),
+            self.measure_filters.clone(),
+            self.measure_order_by.clone(),
+        );
+        (measure_with_source, source)
     }
 
     pub fn is_calculated(&self) -> bool {
@@ -67,9 +92,38 @@ impl MeasureSymbol {
         }
     }
 
-    pub fn evaluate_sql(&self, args: Vec<MemberSqlArg>) -> Result<String, CubeError> {
-        let sql = self.member_sql.call(args)?;
+    pub fn evaluate_sql(
+        &self,
+        visitor: &SqlEvaluatorVisitor,
+        node_processor: Rc<dyn SqlNode>,
+        query_tools: Rc<QueryTools>,
+    ) -> Result<String, CubeError> {
+        let sql = self.member_sql.eval(visitor, node_processor, query_tools)?;
         Ok(sql)
+    }
+
+    pub fn get_dependencies(&self) -> Vec<Rc<MemberSymbol>> {
+        let mut deps = vec![];
+        self.member_sql.extract_symbol_deps(&mut deps);
+        for filter in self.measure_filters.iter() {
+            filter.extract_symbol_deps(&mut deps);
+        }
+        for order in self.measure_order_by.iter() {
+            order.sql_call().extract_symbol_deps(&mut deps);
+        }
+        deps
+    }
+
+    pub fn get_dependent_cubes(&self) -> Vec<String> {
+        let mut cubes = vec![];
+        self.member_sql.extract_cube_deps(&mut cubes);
+        for filter in self.measure_filters.iter() {
+            filter.extract_cube_deps(&mut cubes);
+        }
+        for order in self.measure_order_by.iter() {
+            order.sql_call().extract_cube_deps(&mut cubes);
+        }
+        cubes
     }
 
     pub fn owned_by_cube(&self) -> bool {
@@ -83,7 +137,23 @@ impl MeasureSymbol {
         &self.definition.static_data().measure_type
     }
 
-    pub fn measure_filters(&self) -> &Vec<Rc<EvaluationNode>> {
+    pub fn rolling_window(&self) -> &Option<RollingWindow> {
+        &self.definition.static_data().rolling_window
+    }
+
+    pub fn is_rolling_window(&self) -> bool {
+        self.rolling_window().is_some()
+    }
+
+    pub fn is_running_total(&self) -> bool {
+        self.measure_type() == "runningTotal"
+    }
+
+    pub fn is_cumulative(&self) -> bool {
+        self.is_rolling_window() || self.is_running_total()
+    }
+
+    pub fn measure_filters(&self) -> &Vec<Rc<SqlCall>> {
         &self.measure_filters
     }
 
@@ -102,13 +172,11 @@ impl MeasureSymbol {
     pub fn is_multi_stage(&self) -> bool {
         self.definition.static_data().multi_stage.unwrap_or(false)
     }
-}
 
-impl MemberSymbol for MeasureSymbol {
-    fn cube_name(&self) -> &String {
+    pub fn cube_name(&self) -> &String {
         &self.cube_name
     }
-    fn name(&self) -> &String {
+    pub fn name(&self) -> &String {
         &self.name
     }
 }
@@ -173,11 +241,7 @@ impl SymbolFactory for MeasureSymbolFactory {
         }
     }
 
-    fn build(
-        self,
-        deps: Vec<Dependency>,
-        compiler: &mut Compiler,
-    ) -> Result<Rc<EvaluationNode>, CubeError> {
+    fn build(self, compiler: &mut Compiler) -> Result<Rc<MemberSymbol>, CubeError> {
         let Self {
             cube_name,
             name,
@@ -187,7 +251,7 @@ impl SymbolFactory for MeasureSymbolFactory {
         let mut measure_filters = vec![];
         if let Some(filters) = definition.filters()? {
             for filter in filters.items().iter() {
-                let node = compiler.add_simple_sql_evaluator(cube_name.clone(), filter.sql()?)?;
+                let node = compiler.compile_sql_call(&cube_name, filter.sql()?)?;
                 measure_filters.push(node);
             }
         }
@@ -195,21 +259,19 @@ impl SymbolFactory for MeasureSymbolFactory {
         let mut measure_order_by = vec![];
         if let Some(group_by) = definition.order_by()? {
             for item in group_by.items().iter() {
-                let node = compiler.add_simple_sql_evaluator(cube_name.clone(), item.sql()?)?;
+                let node = compiler.compile_sql_call(&cube_name, item.sql()?)?;
                 measure_order_by.push(MeasureOrderBy::new(node, item.dir()?));
             }
         }
+        let sql = compiler.compile_sql_call(&cube_name, sql)?;
 
-        Ok(EvaluationNode::new_measure(
-            MeasureSymbol::new(
-                cube_name,
-                name,
-                sql,
-                definition,
-                measure_filters,
-                measure_order_by,
-            ),
-            deps,
-        ))
+        Ok(MemberSymbol::new_measure(MeasureSymbol::new(
+            cube_name,
+            name,
+            sql,
+            definition,
+            measure_filters,
+            measure_order_by,
+        )))
     }
 }